requirements.txt
flask
numpy
torch
requests
Pillow
tensorflow